[feat] add feature gate to contexts
