+++ src/main.rs
@@ -179,6 +179,14 @@ async fn process_request(ctx: ContextRef, state: Arc<GlobalState>) {
     }
     let connector = connector.unwrap();
 
+    // Check if connector has requested feature
+    let feature = ctx.read().await.props().request_feature;
+    if !connector.has_feature(feature) {
+        return ctx
+            .on_error(err_msg("unsupported connector feature: {}"))
+            .await;
+    }
+
     ctx.write()
         .await
         .set_state(ContextState::ServerConnecting)
